+++ frontend/src/backoffice/views/Customize/Customize.vue
@@ -4,7 +4,10 @@
 			</p>`" />
 			<div class="app-container">
 			<invoker ref="invoker"></invoker>
-			<search-popup ref="searchPopup" @selected="regionSelected" />
+
+			<search-popup ref="searchPopup" @selected="regionSelected" searchType="r" />
+			<search-popup ref="addMetricPopup" @selected="metricSelected" searchType="m" />
+
 			<relocate @relocated="relocated" ref="Relocate" :useOverlay="false" :useMarker="false"></relocate>
 
 			<div class="md-layout md-gutter">
@@ -47,11 +50,11 @@
 								</div>
 								<div v-if="Startup.ClippingRegionItemId" class="md-layout-item md-size-75 md-small-size-100">
 									<div class="infoRow">
-										{{ Work.StartupExtraInfo.RegionExtraInfo }} &gt; {{ Work.StartupExtraInfo.RegionCaption }}.
+										{{ Work.Startup.RegionExtraInfo }} &gt; {{ Work.Startup.RegionCaption }}.
 									</div>
 								</div>
 								<div class="md-layout-item md-size-100 md-small-size-100">
-									<md-button v-if="Work.CanEdit()" class="md-raised" @click="search">
+									<md-button v-if="Work.CanEdit()" class="md-raised noLeftMargin" @click="search">
 										{{ (Startup.ClippingRegionItemId ? 'Modificar' : 'Seleccionar') }}
 									</md-button>
 								</div>
@@ -78,27 +81,46 @@
 									</div>
 								</div>
 								<div class="md-layout-item md-size-100 md-small-size-100">
-									<md-button v-if="Work.CanEdit()" class="md-raised" @click="relocate">
+									<md-button v-if="Work.CanEdit()" class="md-raised noLeftMargin" @click="relocate">
 										{{ (Startup.Center ? 'Modificar' : 'Indicar') }}
 									</md-button>
 								</div>
 							</div>
 						</md-card-content>
 					</md-card>
-					<md-card v-if="false">
+				</div>
+			</div>
+			<div class="md-layout md-gutter">
+				<div class="md-layout-item md-size-80 md-small-size-100">
+					<md-card>
 						<md-card-header>
-							<div class="md-title">Indicadores
+							<div class="md-title">Indicadores adicionales
 							</div>
 						</md-card-header>
 						<md-card-content>
 							<div class="md-layout md-gutter">
 								<div class="md-layout-item md-size-100 md-small-size-100">
-									Seleccione los indicadores que deben estar activos en el visor al ingresar a la cartografía. Adicionalmente, puede agregar indicadores de otras cartografías al listado de indicadores.
+									En forma complementaria, pueden ofrecerse en el panel superior indicadores con información disponible en el sitio tales como variables demográficas generales o indicadores de infraestructura.
 								</div>
-								<div class="md-layout-item md-size-100 md-small-size-100">
-									<md-button v-if="Work.CanEdit()" class="md-raised">
-										Agregar
+								<div v-if="Work.CanEdit()" class="md-layout-item md-size-100">
+									<md-button @click="addMetric()">
+										<md-icon>add_circle_outline</md-icon>
+									Agregar indicador
 									</md-button>
+								</div>
+								<div class="md-layout-item md-size-80 md-small-size-100">
+									<md-table v-model="Work.ExtraMetrics" md-sort="Caption" md-sort-order="asc" md-card="">
+										<md-table-row slot="md-table-row" slot-scope="{ item }">
+											<md-table-cell md-label="Nombre">{{ item.Caption }}</md-table-cell>
+											<md-table-cell md-label="Acciones" class="mpNoWrap">
+												<div v-if="Work.CanEdit()">
+													<md-button class="md-icon-button" title="Quitar fuente" @click="removeMetric(item)">
+														<md-icon>delete</md-icon>
+													</md-button>
+												</div>
+											</md-table-cell>
+										</md-table-row>
+									</md-table>
 								</div>
 							</div>
 						</md-card-content>
@@ -112,6 +134,7 @@
 import Context from '@/backoffice/classes/Context';
 import f from '@/backoffice/classes/Formatter';
 import str from '@/common/js/str';
+import arr from '@/common/js/arr';
 import Relocate from '@/backoffice/components/Relocate.vue';
 import SearchPopup from '@/backoffice/components/SearchPopup.vue';
 
@@ -139,13 +162,33 @@ export default {
 		Update() {
 			// Valida latitud y longitud.
 			this.$refs.invoker.do(this.Work,
-					this.Work.UpdateStartup);
-			return true;
+				this.Work.UpdateStartup);
+		},
+		metricSelected(metric) {
+			for (var n = 0; n < this.Work.ExtraMetrics.length; n++) {
+				if (this.Work.ExtraMetrics[n].Id === metric.Id) {
+					return;
+				}
+			}
+			var loc = this;
+			this.$refs.invoker.do(this.Work,
+				this.Work.AppendExtraMetric, metric).then(
+					function () {
+						loc.Work.ExtraMetrics.push(metric);
+					});
+		},
+		removeMetric(metric) {
+			var loc = this;
+			this.$refs.invoker.do(this.Work,
+				this.Work.RemoveExtraMetric, metric).then(
+					function () {
+						 arr.Remove(loc.Work.ExtraMetrics, metric);
+					});
 		},
 		regionSelected(item) {
-			this.Startup.ClippingRegionItemId = item.id;
-			this.Work.StartupExtraInfo.RegionCaption = item.caption;
-			this.Work.StartupExtraInfo.RegionExtraInfo = item.extra;
+			this.Startup.ClippingRegionItemId = item.Id;
+			this.Work.Startup.RegionCaption = item.Caption;
+			this.Work.Startup.RegionExtraInfo = item.Extra;
 			if (this.Startup.Type !== 'R') {
 				this.Startup.Type = 'R';
 			}
@@ -153,6 +196,9 @@ export default {
 		},
 		search() {
 			this.$refs.searchPopup.show();
+		},
+		addMetric() {
+			this.$refs.addMetricPopup.show();
 		},
 		relocate() {
 			var lat = null;
@@ -216,6 +262,13 @@ export default {
 .floatRadio {
 	float: left;
   padding-top: 3px!important;
+}
+
+.paddedList {
+	padding: 20px 60px 0px 60px !important;
+}
+.noLeftMargin {
+	margin-left: 0px!important;
 }
 
 .largeOption {