+++ services/src/entities/backoffice/structs/StartupInfo.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace helena\entities\backoffice\structs;
+
+class StartupInfo
+{
+	public $LookupVersion;
+	public $RegionExtraInfo;
+	public $RegionCaption;
+}
+