+++ services/src/entities/backoffice/structs/WorkInfo.php
@@ -9,6 +9,7 @@ class WorkInfo
 	public $Files;
 	public $Permissions;
 	public $Sources;
-	public $StartupExtraInfo;
+	public $ExtraMetrics;
+	public $Startup;
 }